feat(decommission): remove observing-ingester + jetstream-client
tap-ingester is the only firehose-driven ingester remaining; its
Cloud Run service has been the live writer since the verification
window (post #463/#464) confirmed parity with observing-ingester.
Drop the legacy Jetstream-driven path and its sole client crate.

Refactor the absorbed code:

- Files folded into tap-ingester:
  - `database.rs` — methods now take scalar `(did, uri, cid, time, record)`
    instead of a firehose-coupled `&CommitInfo`. Drops cursor get/save
    (Tap manages its own cursor).
  - `error.rs` — drops the `Jetstream` variant; renames `CborDecode`
    → `Decode`.
  - `server.rs` — owns the `ServerState` / `SharedState` types only;
    HTTP handlers and dashboard HTML stay in `dashboard.rs`.
  - `types.rs` — keeps `IngesterStats`, `RecentEvent`, and the five
    collection constants. Drops `IngesterConfig`, `ALL_COLLECTIONS`,
    and `resolve_collection_names` (Tap-driven config doesn't need
    them).
  - `media_resolver.rs` — unchanged.

- Worker bins relocated to their own crates:
  - `crates/observing-backfill/` — one-shot CLI for seeding records
    from a PDS via `listRecords`. CI's e2e and post-incident runs
    invoke this directly.
  - `crates/observing-resolve-taxa/` — background worker that fills
    `accepted_taxon_key` on identifications. Operated independently
    of the ingester.

- Infra:
  - `Dockerfile`: drop the `runtime-observing-ingester` stage and
    the `observing-ingester` SERVICE comment.
  - `.github/workflows/ci.yml`: drop `observing-ingester` from the
    build matrix and the `Deploy ingester` step. e2e now builds Tap
    from the pinned indigo commit (same as the production Dockerfile)
    and starts `tap-ingester` instead. `cargo sqlx prepare --check`
    moves to `--workspace` mode (single cache at the repo root).
  - `process-compose.yaml`: replace the `ingester` process with a
    `tap-ingester` process pointing at a local sqlite tap.db.

The cross-repo subject resolver from #464 is intentionally not
included — that PR remains independent.
