+++ .github/workflows/ci.yml
@@ -250,8 +250,7 @@ jobs:
           tool: sqlx-cli
           no-default-features: true
           features: postgres
-      - working-directory: crates/observing-ingester
-        run: cargo sqlx prepare --check
+      - run: cargo sqlx prepare --workspace --check
 
   rust-lexicons-check:
     runs-on: ubuntu-latest
@@ -303,7 +302,7 @@ jobs:
       - name: Run database migrations
         run: for f in crates/observing-db/migrations/*.sql; do psql $DATABASE_URL -f "$f"; done
       - name: Backfill from PDS
-        run: cargo run --bin backfill -- did:plc:jh6n3ntljfhhtr4jbvrm3k5b
+        run: cargo run -p observing-backfill -- did:plc:jh6n3ntljfhhtr4jbvrm3k5b
 
   # End-to-end tests
   e2e:
@@ -346,14 +345,32 @@ jobs:
         run: npm run build
       - name: Build appview
         run: cargo build -p observing-appview
-      - name: Build ingester
-        run: cargo build -p observing-ingester --bin observing-ingester
+      - name: Build tap-ingester
+        run: cargo build -p tap-ingester --bin tap-ingester
+      - uses: actions/setup-go@v5
+        with:
+          go-version: '1.26'
+      - name: Build and install upstream `tap`
+        # Same indigo commit the production Dockerfile pins.
+        env:
+          INDIGO_REV: ce62b8fce9e01434213a69cb251852b2c9436cb9
+        run: |
+          mkdir -p /tmp/indigo && cd /tmp/indigo
+          git init . >/dev/null
+          git remote add origin https://github.com/bluesky-social/indigo.git
+          git fetch --depth 1 origin "$INDIGO_REV"
+          git checkout FETCH_HEAD
+          GODEBUG=netdns=go go build -tags timetzdata -o /usr/local/bin/tap ./cmd/tap
       - name: Install Playwright
         run: npx playwright install --with-deps chromium
       - name: Start services
+        env:
+          # tap-ingester's default is sqlite:///data/tap.db (a Cloud Run
+          # container path); point at a local file in CI instead.
+          TAP_DATABASE_URL: sqlite:/tmp/tap.db
         run: |
           PORT=3000 ./target/debug/observing-appview > /tmp/appview.log 2>&1 &
-          PORT=8090 ./target/debug/observing-ingester > /tmp/ingester.log 2>&1 &
+          PORT=8090 ./target/debug/tap-ingester > /tmp/tap-ingester.log 2>&1 &
       - name: Wait for services
         run: |
           for i in $(seq 1 60); do
@@ -374,8 +391,8 @@ jobs:
           echo "===== appview log ====="
           tail -200 /tmp/appview.log || true
           echo
-          echo "===== ingester log ====="
-          tail -200 /tmp/ingester.log || true
+          echo "===== tap-ingester log ====="
+          tail -200 /tmp/tap-ingester.log || true
       - uses: actions/upload-artifact@v4
         if: always()
         with:
@@ -394,7 +411,6 @@ jobs:
       matrix:
         service:
           - observing-appview
-          - observing-ingester
           - observing-species-id
           - observing-migrate
           - tap-ingester
@@ -483,25 +499,6 @@ jobs:
             --region=$REGION \
             --wait
 
-      - name: Deploy ingester
-        run: |
-          gcloud run deploy observing-ingester \
-            --image=$REGISTRY/observing-ingester:latest \
-            --region=$REGION \
-            --allow-unauthenticated \
-            --cpu=1 \
-            --memory=2Gi \
-            --add-cloudsql-instances=$PROJECT_ID:$REGION:observing-db \
-            --set-env-vars=RUST_LOG=observing_ingester=info,DB_HOST=/cloudsql/$PROJECT_ID:$REGION:observing-db,DB_NAME=observing,DB_USER=ingester_runtime,JETSTREAM_URL=wss://jetstream2.us-east.bsky.network/subscribe \
-            --set-secrets=DB_PASSWORD=observing-db-ingester-password:latest \
-            --min-instances=1 \
-            --max-instances=1 \
-            --timeout=3600
-
-      # Side-by-side with observing-ingester during the verification window.
-      # Bundles its own `tap` binary spawned as a child process; SQLite state
-      # at /data/tap.db is instance-ephemeral (Tap re-discovers + re-backfills
-      # on instance restart, which is fine for current scale).
       - name: Deploy tap-ingester
         run: |
           gcloud run deploy tap-ingester \
