--- .sqlx/query-bc08c5a66cbbed9834a460ed7b6b4471ad64fa416dcd24931717be33b5c73bb4.json
@@ -1,20 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "SELECT value FROM ingester_state WHERE key = 'cursor'",
-  "describe": {
-    "columns": [
-      {
-        "ordinal": 0,
-        "name": "value",
-        "type_info": "Text"
-      }
-    ],
-    "parameters": {
-      "Left": []
-    },
-    "nullable": [
-      false
-    ]
-  },
-  "hash": "bc08c5a66cbbed9834a460ed7b6b4471ad64fa416dcd24931717be33b5c73bb4"
-}
