--- crates/jetstream-client/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "jetstream-client"
-version = "0.1.0"
-edition = "2021"
-description = "AT Protocol Jetstream WebSocket client"
-license = "MIT OR Apache-2.0"
-
-[dependencies]
-# Async runtime
-tokio = { workspace = true }
-
-# WebSocket
-tokio-tungstenite = { version = "0.29", features = ["native-tls"] }
-futures-util = { workspace = true }
-
-# Serialization
-serde = { workspace = true }
-serde_json = { workspace = true }
-
-# Logging
-tracing = { workspace = true }
-
-# Time
-chrono = { workspace = true }
