+++ crates/observing-backfill/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "observing-backfill"
+version = "0.1.0"
+edition = "2021"
+description = "One-shot CLI for seeding records from a PDS via listRecords"
+license = "MIT OR Apache-2.0"
+
+[dependencies]
+atproto-identity = { path = "../atproto-identity" }
+observing-db = { path = "../observing-db", features = ["processing"] }
+
+tokio = { workspace = true }
+sqlx = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+reqwest = { workspace = true }
+chrono = { workspace = true }
+clap = { version = "4", features = ["derive"] }
+tracing = { workspace = true }
+tracing-subscriber = { workspace = true }
+
+[[bin]]
+name = "observing-backfill"
+path = "src/main.rs"
