--- crates/observing-ingester/.sqlx/query-3796d4d939bcea23d8dd3da763b399a888451b81ad47f47098ed7464ffdded36.json
@@ -1,14 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "DELETE FROM comments WHERE uri = $1",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Text"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "3796d4d939bcea23d8dd3da763b399a888451b81ad47f47098ed7464ffdded36"
-}
