--- crates/observing-ingester/.sqlx/query-85a846cea136f7cea8b49408609074a85f9a7dc06dc74927addd50631139c15d.json
@@ -1,22 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n            INSERT INTO comments (\n                uri, did, cid, subject_uri, subject_cid, body,\n                reply_to_uri, reply_to_cid, created_at, indexed_at\n            )\n            VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, NOW())\n            ON CONFLICT (uri) DO UPDATE SET\n                cid = EXCLUDED.cid,\n                subject_uri = EXCLUDED.subject_uri,\n                subject_cid = EXCLUDED.subject_cid,\n                body = EXCLUDED.body,\n                reply_to_uri = EXCLUDED.reply_to_uri,\n                reply_to_cid = EXCLUDED.reply_to_cid,\n                created_at = EXCLUDED.created_at,\n                indexed_at = NOW()\n            ",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Text",
-        "Text",
-        "Text",
-        "Text",
-        "Text",
-        "Text",
-        "Text",
-        "Text",
-        "Timestamp"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "85a846cea136f7cea8b49408609074a85f9a7dc06dc74927addd50631139c15d"
-}
