--- crates/observing-ingester/.sqlx/query-afd117401141b4ae2df63650cba7470495a854136e20527d1059fdaf319190b3.json
@@ -1,14 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "DELETE FROM identifications WHERE uri = $1",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Text"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "afd117401141b4ae2df63650cba7470495a854136e20527d1059fdaf319190b3"
-}
