--- crates/observing-ingester/.sqlx/query-c2908c087344c6b6bb0c6ace5eea245f539555b7b77ec12fb5a3bed82263d3ca.json
@@ -1,14 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "DELETE FROM occurrences WHERE uri = $1",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Text"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "c2908c087344c6b6bb0c6ace5eea245f539555b7b77ec12fb5a3bed82263d3ca"
-}
