--- crates/observing-ingester/Cargo.toml
@@ -1,67 +1,0 @@
-[package]
-name = "observing-ingester"
-version = "0.1.0"
-edition = "2021"
-description = "High-performance AT Protocol firehose ingester for Observ.ing"
-license = "MIT OR Apache-2.0"
-
-[dependencies]
-# Async runtime
-tokio = { workspace = true }
-
-# Jetstream client
-jetstream-client = { path = "../jetstream-client" }
-
-# DID newtype
-atproto-identity = { path = "../atproto-identity" }
-
-# AT URI parsing (for resolving associatedMedia strong refs)
-at-uri-parser = { path = "../at-uri-parser" }
-
-# DID → PDS resolution (for fetching media records)
-atproto-blob-resolver = { path = "../atproto-blob-resolver" }
-
-# Database
-sqlx = { workspace = true }
-observing-db = { path = "../observing-db", features = ["processing"] }
-observing-lexicons = { path = "../observing-lexicons" }
-
-# Taxonomy resolver upstream (GBIF live API → local `taxa` cache)
-observing-taxonomy-gbif = { path = "../observing-taxonomy-gbif" }
-
-# HTTP server
-axum = { workspace = true }
-tower-http = { workspace = true }
-
-# Serialization
-serde = { workspace = true }
-serde_json = { workspace = true }
-
-# Logging
-tracing = { workspace = true }
-tracing-subscriber = { workspace = true }
-tracing-stackdriver = { workspace = true }
-
-# Time
-chrono = { workspace = true }
-
-# CLI
-clap = { version = "4", features = ["derive"] }
-
-# HTTP client (for backfill)
-reqwest = { workspace = true }
-
-# URL parsing
-url = { workspace = true }
-urlencoding = { workspace = true }
-
-[dev-dependencies]
-tower = { workspace = true, features = ["util"] }
-
-[lib]
-name = "observing_ingester"
-path = "src/lib.rs"
-
-[[bin]]
-name = "observing-ingester"
-path = "src/main.rs"
